+++ docs/changes.rst
@@ -5,6 +5,8 @@ Weblate 5.12
 
 .. rubric:: New features
 
+* Added :http:post:`/api/units/(int:id)/comments/` to create a new comment for the given translation unit.
+
 .. rubric:: Improvements
 
 * :ref:`search-strings` supports ``has:location`` lookup.